+++ packages/codemod/README.md
@@ -78,18 +78,18 @@ import components and types from `antd` to `@oceanbase/design` and `@oceanbase/u
   export default Demo;
 ```
 
-### `page-container-to-oceanbase-ui`
+### `techui-and-pro-components-to-oceanbase-ui`
 
-import `PageContainer` from `@alipay/tech-ui` to `@ant-design/pro-components` and `@oceanbase/ui`.
+import `PageContainer` from `@alipay/tech-ui` and `@ant-design/pro-components` to `@oceanbase/ui`.
 
 ```diff
   import React from 'react';
 - import { PageContainer, ProTable, WelcomeCard } from '@alipay/tech-ui';
 - import type { PageContainerProps, ProTableProps, WelcomeCardProps } from '@alipay/tech-ui';
-+ import { ProTable, WelcomeCard } from '@alipay/tech-ui';
-+ import { PageContainer } from '@oceanbase/ui';
-+ import type { ProTableProps, WelcomeCardProps } from '@alipay/tech-ui';
-+ import type { PageContainerProps } from '@oceanbase/ui';
++ import { WelcomeCard } from '@alipay/tech-ui';
++ import { PageContainer, ProTable } from '@oceanbase/ui';
++ import type { WelcomeCardProps } from '@alipay/tech-ui';
++ import type { PageContainerProps, ProTableProps } from '@oceanbase/ui';
 
   const Demo = () => {
     return <div />;
