--- packages/codemod/transforms/__testfixtures__/page-container-to-oceanbase-ui/pro-components.input.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { PageContainer, ProTable } from '@ant-design/pro-components';
-import type { PageContainerProps, ProTableProps } from '@ant-design/pro-components';
-
-const Demo = () => {
-  return <div />;
-};
-
-export default Demo;
