--- packages/codemod/transforms/__testfixtures__/page-container-to-oceanbase-ui/pro-components.output.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { ProTable } from '@ant-design/pro-components';
-import { PageContainer } from '@oceanbase/ui';
-import type { ProTableProps } from '@ant-design/pro-components';
-
-import type { PageContainerProps } from '@oceanbase/ui';
-
-const Demo = () => {
-  return <div />;
-};
-
-export default Demo;
