--- packages/codemod/transforms/__testfixtures__/page-container-to-oceanbase-ui/techui.input.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { PageContainer, ProTable, WelcomeCard } from '@alipay/tech-ui';
-import type { PageContainerProps, ProTableProps, WelcomeCardProps } from '@alipay/tech-ui';
-
-const Demo = () => {
-  return <div />;
-};
-
-export default Demo;
