--- packages/codemod/transforms/__testfixtures__/page-container-to-oceanbase-ui/techui.output.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { ProTable, WelcomeCard } from '@alipay/tech-ui';
-import { PageContainer } from '@oceanbase/ui';
-import type { ProTableProps, WelcomeCardProps } from '@alipay/tech-ui';
-
-import type { PageContainerProps } from '@oceanbase/ui';
-
-const Demo = () => {
-  return <div />;
-};
-
-export default Demo;
