+++ packages/codemod/transforms/__testfixtures__/techui-and-pro-components-to-oceanbase-ui/pro-components.input.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { LightFilter, PageContainer, ProCard, ProTable } from '@ant-design/pro-components';
+import type { LightFilterProps, PageContainerProps, ProCardProps, ProTableProps } from '@ant-design/pro-components';
+
+const Demo = () => {
+  return <div />;
+};
+
+export default Demo;
