+++ packages/codemod/transforms/__testfixtures__/techui-and-pro-components-to-oceanbase-ui/pro-components.output.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { LightFilter, PageContainer, ProCard, ProTable } from '@oceanbase/ui';
+import type { LightFilterProps, PageContainerProps, ProCardProps, ProTableProps } from '@oceanbase/ui';
+
+const Demo = () => {
+  return <div />;
+};
+
+export default Demo;
