+++ packages/codemod/transforms/__testfixtures__/techui-and-pro-components-to-oceanbase-ui/techui.input.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { LightFilter, PageContainer, ProCard, ProTable, WelcomeCard } from '@alipay/tech-ui';
+import type { LightFilterProps, PageContainerProps, ProCardProps, ProTableProps, WelcomeCardProps } from '@alipay/tech-ui';
+
+const Demo = () => {
+  return <div />;
+};
+
+export default Demo;
