+++ packages/codemod/transforms/__testfixtures__/techui-and-pro-components-to-oceanbase-ui/techui.output.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { WelcomeCard } from '@alipay/tech-ui';
+import { LightFilter, PageContainer, ProCard, ProTable } from '@oceanbase/ui';
+import type { WelcomeCardProps } from '@alipay/tech-ui';
+
+import type { LightFilterProps, PageContainerProps, ProCardProps, ProTableProps } from '@oceanbase/ui';
+
+const Demo = () => {
+  return <div />;
+};
+
+export default Demo;
