+++ packages/codemod/transforms/obui-to-oceanbase-design-and-ui.js
@@ -1,4 +1,5 @@
 const importComponent = require('./utils/import-component');
+const { proComponents } = require('./utils/config');
 
 module.exports = (file, api, options) => {
   return importComponent(file, api, {
@@ -32,46 +33,7 @@ module.exports = (file, api, options) => {
           'TreeSearch',
           'Welcome',
           /* pro-components */
-          'PageContainer',
-          'ProLayout',
-          'ProCard',
-          'WaterMark',
-          'StatisticCard',
-          'CheckCard',
-          'ProFormText',
-          'ProFormDigit',
-          'ProFormTextArea',
-          'ProFormCaptcha',
-          'ProFormDatePicker',
-          'ProFormDateTimePicker',
-          'ProFormDateRangePicker',
-          'ProFormDateTimeRangePicker',
-          'ProFormSelect',
-          'ProFormTreeSelect',
-          'ProFormCheckbox',
-          'ProFormRadio',
-          'ProFormSlider',
-          'ProFormSwitch',
-          'ProFormUploadButton',
-          'ProFormUploadDragger',
-          'ProFormMoney',
-          'ProFormSegmented',
-          'ProFormList',
-          'BetaSchemaForm',
-          'QueryFilter',
-          'LightFilter',
-          'StepsForm',
-          'ModalForm',
-          'DrawerForm',
-          'LoginForm',
-          'LoginFormPage',
-          'ProTable',
-          'EditableProTable',
-          'DragSortTable',
-          'ProList',
-          'ProDescriptions',
-          'ProSkeleton',
-          'ProField',
+          ...proComponents.components,
         ],
         types: [
           'ActionProps',
@@ -106,46 +68,7 @@ module.exports = (file, api, options) => {
           'Node',
           'WelcomeProps',
           /* pro-components */
-          'PageContainerProps',
-          'ProLayoutProps',
-          'ProCardProps',
-          'WaterMarkProps',
-          'StatisticCardProps',
-          'CheckCardProps',
-          'ProFormTextProps',
-          'ProFormDigitProps',
-          'ProFormTextAreaProps',
-          'ProFormCaptchaProps',
-          'ProFormDatePickerProps',
-          'ProFormDateTimePickerProps',
-          'ProFormDateRangePickerProps',
-          'ProFormDateTimeRangePickerProps',
-          'ProFormSelectProps',
-          'ProFormTreeSelectProps',
-          'ProFormCheckboxProps',
-          'ProFormRadioProps',
-          'ProFormSliderProps',
-          'ProFormSwitchProps',
-          'ProFormUploadButtonProps',
-          'ProFormUploadDraggerProps',
-          'ProFormMoneyProps',
-          'ProFormSegmentedProps',
-          'ProFormListProps',
-          'BetaSchemaFormProps',
-          'QueryFilterProps',
-          'LightFilterProps',
-          'StepsFormProps',
-          'ModalFormProps',
-          'DrawerFormProps',
-          'LoginFormProps',
-          'LoginFormPageProps',
-          'ProTableProps',
-          'EditableProTableProps',
-          'DragSortTableProps',
-          'ProListProps',
-          'ProDescriptionsProps',
-          'ProSkeletonProps',
-          'ProFieldProps',
+          ...proComponents.types,
         ],
         paths: ['/locale/', '/locale/'],
       },
