+++ packages/codemod/transforms/utils/config.js
@@ -4,6 +4,94 @@ const printOptions = {
   wrapColumn: 500,
 };
 
+const proComponents = {
+  components: [
+    'PageContainer',
+    'ProLayout',
+    'ProCard',
+    'WaterMark',
+    'StatisticCard',
+    'CheckCard',
+    'ProFormText',
+    'ProFormDigit',
+    'ProFormTextArea',
+    'ProFormCaptcha',
+    'ProFormDatePicker',
+    'ProFormDateTimePicker',
+    'ProFormDateRangePicker',
+    'ProFormDateTimeRangePicker',
+    'ProFormSelect',
+    'ProFormTreeSelect',
+    'ProFormCheckbox',
+    'ProFormRadio',
+    'ProFormSlider',
+    'ProFormSwitch',
+    'ProFormUploadButton',
+    'ProFormUploadDragger',
+    'ProFormMoney',
+    'ProFormSegmented',
+    'ProFormList',
+    'BetaSchemaForm',
+    'QueryFilter',
+    'LightFilter',
+    'StepsForm',
+    'ModalForm',
+    'DrawerForm',
+    'LoginForm',
+    'LoginFormPage',
+    'ProTable',
+    'EditableProTable',
+    'DragSortTable',
+    'ProList',
+    'ProDescriptions',
+    'ProSkeleton',
+    'ProField',
+  ],
+  types: [
+    'PageContainerProps',
+    'ProLayoutProps',
+    'ProCardProps',
+    'WaterMarkProps',
+    'StatisticCardProps',
+    'CheckCardProps',
+    'ProFormTextProps',
+    'ProFormDigitProps',
+    'ProFormTextAreaProps',
+    'ProFormCaptchaProps',
+    'ProFormDatePickerProps',
+    'ProFormDateTimePickerProps',
+    'ProFormDateRangePickerProps',
+    'ProFormDateTimeRangePickerProps',
+    'ProFormSelectProps',
+    'ProFormTreeSelectProps',
+    'ProFormCheckboxProps',
+    'ProFormRadioProps',
+    'ProFormSliderProps',
+    'ProFormSwitchProps',
+    'ProFormUploadButtonProps',
+    'ProFormUploadDraggerProps',
+    'ProFormMoneyProps',
+    'ProFormSegmentedProps',
+    'ProFormListProps',
+    'BetaSchemaFormProps',
+    'QueryFilterProps',
+    'LightFilterProps',
+    'StepsFormProps',
+    'ModalFormProps',
+    'DrawerFormProps',
+    'LoginFormProps',
+    'LoginFormPageProps',
+    'ProTableProps',
+    'EditableProTableProps',
+    'DragSortTableProps',
+    'ProListProps',
+    'ProDescriptionsProps',
+    'ProSkeletonProps',
+    'ProFieldProps',
+  ],
+};
+
 module.exports = {
   printOptions,
+  proComponents,
 };
